requirements.txt
python-telegram-bot==20.7
pytz==2023.3
httpx[socks]==0.25.2